requirements.txt
torch
torchvision
pytorch-ignite
tqdm
lmdb
tensorboard